@@ -22,14 +22,18 @@ type Silence struct {
 
 // +k8s:openapi-gen=true
 type SilenceSpec struct {
-	TargetTags []TargetTag `json:"targetTags"`
+	TargetTags []TargetTag `json:"targetTags,omitempty"`
 	Matchers   []Matcher   `json:"matchers"`
 
 	// Owner is GitHub username of a person who created and/or owns the silence.
-	Owner *string `json:"owner,omitempty"`
+	Owner string `json:"owner,omitempty"`
 
 	// PostmortemURL is a link to a document describing the problem.
+	// Deprecated: Use IssueURL instead.
 	PostmortemURL *string `json:"postmortem_url,omitempty"`
+
+	// IssueURL is a link to a GitHub issue describing the problem.
+	IssueURL string `json:"issue_url,omitempty"`
 }
 
 type TargetTag struct {
